Simplify event handling by only handling one event at a time
Previously the code added each event to a list (well, really a map)
which was maintained across iterations as it propagated 'down' through
the MSF; each event was removed from the list as soon as it was handled.
However, this meant that if the MSF at some point *stopped* listening to
an event, the list would accumulate a backlog of events, with
unintuitive consequences. With this commit, the MSF instead handles only
a single event per iteration, which is discarded if unhandled (and
errors if handled twice, since that case probably indicates a bug in the
code).